chore: move required contacts validator
