+++ server/validation/utils.ts
@@ -1,9 +1,6 @@
-import { ContactType } from "@prisma/client";
-import { array, coerce, date, define, refine, string } from "superstruct";
+import { coerce, date, define, string } from "superstruct";
 
 import validator from "validator";
-
-import { ContactCreateImplicitValidator } from "./contact";
 
 /**
  * Superstruct validator for an uuid.
@@ -24,23 +21,3 @@ export const uuid = (version: 3 | 4 | 5 | "all" = 4) =>
  */
 export const dateString = () =>
   coerce(date(), string(), (date) => new Date(date));
-
-/**
- * Makes sure that a list of contacts includes the required items.
- * Useful when creating a new object where contacts are used.
- */
-export const requiredContact = refine(
-  array(ContactCreateImplicitValidator),
-  "requiredContacts",
-  (contacts) => {
-    if (!contacts.find((c) => c.type === ContactType.EMAIL)) {
-      return "Contacts require an email address.";
-    }
-
-    if (!contacts.find((c) => c.type === ContactType.PHONE)) {
-      return "Contacts require an phone number.";
-    }
-
-    return true;
-  },
-);
